Improved reusability
