+++ extensions/rss-reader/src/stories.tsx
@@ -1,12 +1,14 @@
-import { ActionPanel, Color, CopyToClipboardAction, getLocalStorageItem, Icon, ImageLike, List, OpenInBrowserAction, randomId, setLocalStorageItem, showToast, ToastStyle } from "@raycast/api"
-import { Feed } from "./feeds"
+import { ActionPanel, Color, CopyToClipboardAction, getLocalStorageItem, Icon, ImageLike, List, OpenInBrowserAction, PushAction, randomId, setLocalStorageItem, showToast, ToastStyle } from "@raycast/api"
+import { Feed, getFeeds } from "./feeds"
+import AddFeedForm from "./subscription-form"
 import Parser from "rss-parser"
 import TimeAgo from 'javascript-time-ago'
 import en from 'javascript-time-ago/locale/en.json'
+import { useEffect, useState } from "react"
 
 const parser = new Parser({});
 
-export interface Story {
+interface Story {
   guid: string,
   title: string,
   link?: string,
@@ -23,7 +25,7 @@ type FeedLastViewed = {
 TimeAgo.addDefaultLocale(en)
 const timeAgo = new TimeAgo('en-US')
 
-export function StoryListItem(props: { item: Story }) {
+function StoryListItem(props: { item: Story }) {
   return (
     <List.Item
       title={ props.item.title }
@@ -54,7 +56,7 @@ function Actions(props: { item: Story }) {
   );
 }
 
-export function ItemToStory(item: Parser.Item, feed: Feed, lastViewed: number) {
+function ItemToStory(item: Parser.Item, feed: Feed, lastViewed: number) {
   return {
     guid: item.guid || randomId(),
     title: item.title || "No title",
@@ -66,7 +68,7 @@ export function ItemToStory(item: Parser.Item, feed: Feed, lastViewed: number) {
   } as Story
 }
 
-export async function getStories(feeds: Feed[]) {
+async function getStories(feeds: Feed[]) {
   const feedLastViewedString = await getLocalStorageItem("feedLastViewed") as string
   const feedLastViewed = feedLastViewedString ? JSON.parse(feedLastViewedString) as FeedLastViewed : {} as FeedLastViewed
 
@@ -90,3 +92,56 @@ export async function getStories(feeds: Feed[]) {
   await setLocalStorageItem("feedLastViewed", JSON.stringify(feedLastViewed))
   return storyItems
 }
+
+export function StoriesList(props: { feeds? : Feed[] }) {
+  const [feeds, setFeeds] = useState<Feed[]>([] as Feed[])
+  const [stories, setStories] = useState<Story[]>([] as Story[])
+  const [loading, setLoading] = useState(false);
+
+  async function fetchFeeds() {
+    if (props?.feeds) {
+      setFeeds(props.feeds)
+    } else {
+      setFeeds(await getFeeds())
+    }
+  } 
+
+  async function fetchStories() {
+    if (feeds.length === 0) {
+      return
+    }
+    setLoading(true)
+    setStories(await getStories(feeds))
+    setLoading(false)
+  }
+
+  useEffect(() => {
+    fetchFeeds()
+  }, [])
+
+  useEffect(() => {
+    fetchStories()
+  }, [feeds])
+
+  return (
+    <List 
+      isLoading = { loading }
+      actions={ (!props?.feeds) && (
+        <ActionPanel>
+          <PushAction
+            title="Add Feed"
+            target={ <AddFeedForm callback={ setFeeds } /> }
+            icon={{ source: Icon.Plus, tintColor: Color.Green }}
+            shortcut={{ modifiers: ["cmd"], key: "n" }}
+          />
+        </ActionPanel>
+      )}
+    >
+      {stories.map((story) => (
+        <StoryListItem key={ story.guid } item={ story } />
+      ))}
+    </List>
+  )
+}
+
+export default () => { return (<StoriesList />) }
